Remove duplicate jar and change POM to match com.forter.* Also move clojure/lang to aphyr/riemann/client
