+++ src/main/java/com/aphyr/riemann/client/IPromise.java
@@ -2,7 +2,6 @@ package com.aphyr.riemann.client;
 
 import java.util.concurrent.TimeUnit;
 import java.io.IOException;
-import clojure.lang.IDeref;
 
 public interface IPromise<T> extends IDeref {
   public void deliver(Object value);
